pyproject.toml
[project]
name = "spectacular-dmr-adapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.11",
    "django-extensions>=4.1",
    "django-modern-rest[msgspec,pydantic]",
    "djangorestframework>=3.16.1",
    "drf-spectacular[sidecar]>=0.29.0",
    "pyjwt>=2.11.0",
    "xmltodict>=1.0.3",
]

[dependency-groups]
dev = [
    "ipython>=9.10.0",
]

[tool.uv.sources]
django-modern-rest = { git = "https://github.com/wemake-services/django-modern-rest", rev = "master" }
